Correctly handle deleting the ship or loading a new one in the editor
Removal of vacationing kerbals from a ship's root part is handled by
keeping track of whether the root part has been cleaned, and doing it
again if the root part is replaced later.  But the implementation was
only looking at onEditorShipModified events, to check whether a root
part exists, and that's not enough:

 * onEditorShipModified doesn't fire when the root is removed, either by
   picking it up and dropping it back into the parts list or by clicking
   the editor's "new" button.  Consequently rootExists would't change
   from true to false, so cleanedRoot wouldn't change from true to
   false, so CleanManifest() wouldn't be called when a new root part was
   placed later.
 * Loading a new ship design does fire onEditorShipModified, but if the
   old ship design had a root part and the new one did too, that handler
   wouldn't notice that anything had changed, so CleanManifest()
   wouldn't be run on the new ship.

I've added handlers for the onEditorPodDeleted and onEditorReset events
to "notice" when the root part has been removed, and a handler for
onEditorLoaded to "notice" when the root part has been replaced with
that of a different ship design.  This ensures that CleanManifest()
always runs when the root part changes, so it doesn't get vacationing
kerbals assigned to it as crew.