Listen for onEditorPodPicked instead of onEditorShipModified
The only "modification" we're interested in is when a root part is
placed, and there's an event specifically for that, which is more
appropriate than the event for when *anything* changes.  This also
removes the need for the CheckRoot function, since the event itself
indicates that there's a root part.